# src/main/java/com/customermanagement/beans/AdditionalAttributesRequestBean.java
package com.customermanagement.beans;

public class AdditionalAttributesRequestBean {

	private String additionalAttributeName;
	
	private String additionalAttributeValue;
	
	public String getAdditionalAttributeName() {
		return additionalAttributeName;
	}

	public void setAdditionalAttributeName(String additionalAttributeName) {
		this.additionalAttributeName = additionalAttributeName;
	}

	public String getAdditionalAttributeValue() {
		return additionalAttributeValue;
	}

	public void setAdditionalAttributeValue(String additionalAttributeValue) {
		this.additionalAttributeValue = additionalAttributeValue;
	}

}